OVU-153 Moves the Save & Close button to the dialog header
