+++ components/dialog/dialog.jsx
@@ -2,15 +2,22 @@ import React, { useRef } from 'react';
 import { createPortal } from 'react-dom';
 import PropTypes from 'prop-types';
 import classnames from 'classnames';
+import saveState from 'services/saveState';
+import insertEmbed from 'services/insertEmbed';
+import { displayDismissableAlert } from 'services/alert';
 import portalId from 'services/portalId';
 import withTrappedTabs from 'services/withTrappedTabs';
+import getPostAttribute from 'services/getPostAttribute';
 import eventBus from 'services/eventBus';
 import postIsEmpty from 'services/postIsEmpty';
+import OovvuuDataContext from 'components/app/context';
+import ActionButton from 'components/shared/actionButton';
 import LoadingWrapper from 'components/shared/loading';
 import buttons from 'components/shared/actionButton/actionButton.scss';
 import OovvuuSVGLogo from 'assets/oovvuu-logo.svg';
 import WPVIPSVGLogo from 'assets/wp-vip-logo.svg';
 import CloseSVG from 'assets/close.svg';
+import SaveSVG from 'assets/save.svg';
 import styles from './dialog.scss';
 
 /**
@@ -24,12 +31,81 @@ const Dialog = ({
   trappedTabsRoot,
 }) => {
   const { __ } = wp.i18n;
+  const {
+    state,
+    state: {
+      isPositionTwoEnabled,
+      selectedVideos: {
+        positionTwo,
+      },
+    },
+    dispatch,
+  } = React.useContext(OovvuuDataContext);
 
   /**
    * Create references to elements.
    */
   const closeButtonRef = useRef(null);
   const backToTopButtonRef = useRef(null);
+
+  /**
+   * Handles the save action when a user clicks the save button.
+   */
+  const handleSave = async () => {
+    dispatch({ type: 'SET_LOADING_STATE' });
+
+    // Note: Loading state is cleared within the `saveState` service.
+    const response = await saveState(state, getPostAttribute('id'));
+    const {
+      hasError,
+      data,
+      error: {
+        message,
+      } = {},
+    } = response;
+
+    if (!hasError) {
+      // Embed id.
+      const positionTwoEmbedId = data?.embeds?.positionTwo?.id;
+
+      // Insert a new Oovvuu embed to the editor.
+      insertEmbed(positionTwoEmbedId, positionTwo, isPositionTwoEnabled);
+
+      /**
+       * saveState() returns updated state, with flag that data has been loaded
+       *   from meta. This needs to be sync'd back to state.
+       */
+      dispatch({ type: 'RESET_STATE', payload: data });
+
+      // Close the Dialog.
+      closeDialog();
+    } else {
+      dispatch({ type: 'CLEAR_LOADING_STATE' });
+
+      displayDismissableAlert({ message });
+    }
+  };
+
+  /**
+   * The Oovvuu for WordPress branding.
+   */
+  const Branding = () => (
+    <span className={styles.branding}>
+      <span className={styles.oovvuuLogo}>
+        <OovvuuSVGLogo />
+      </span>
+      <span
+        className={styles.logoSeparator}
+        aria-hidden="true"
+      >
+        {__('for', 'oovvuu')}
+      </span>
+      <span className={styles.wpVipLogo}>
+        <WPVIPSVGLogo />
+      </span>
+      <h1 className="screen-reader-only">{__('Oovvuu for WordPress', 'oovvuu')}</h1>
+    </span>
+  );
 
   /**
    * Conditionally sets focus on modal open to either the close button or the
@@ -63,13 +139,16 @@ const Dialog = ({
           aria-hidden={isLoading}
         >
           <div className={styles.header}>
-            <span className={styles.oovvuuLogo}>
-              <OovvuuSVGLogo />
-            </span>
-            <span className={styles.logoSeparator}>{__('for', 'oovvuu')}</span>
-            <span className={styles.wpVipLogo}>
-              <WPVIPSVGLogo />
-            </span>
+            <Branding />
+
+            <ActionButton
+              className={styles.saveButton}
+              buttonStyle="primary"
+              onClickHandler={handleSave}
+            >
+              <SaveSVG />
+              <>{__('Save and Close', 'oovvuu')}</>
+            </ActionButton>
 
             <button
               ref={closeButtonRef}
