+++ components/dialog/index.jsx
@@ -1,12 +1,9 @@
 import React from 'react';
 import ActionButton from 'components/shared/actionButton';
 import getPostAttribute from 'services/getPostAttribute';
-import saveState from 'services/saveState';
 import OovvuuDataContext from 'components/app/context';
-import SaveSVG from 'assets/save.svg';
 import OovvuuSmallSVGLogo from 'assets/oovvuu-small-logo.svg';
-import insertEmbed from 'services/insertEmbed';
-import { confirmThenProceed, displayDismissableAlert } from 'services/alert';
+import { confirmThenProceed } from 'services/alert';
 import TopicsPanelWrapper from './topicsPanel';
 import PositionsPanelWrapper from './positionsPanel';
 import KeywordPanel from './keywordPanel';
@@ -20,14 +17,9 @@ const DialogWrapper = () => {
   const { i18n: { __ } } = wp;
   const [isOpen, setIsOpen] = React.useState(false);
   const {
-    state,
     state: {
       isLoadedFromMeta,
       isLoading,
-      isPositionTwoEnabled,
-      selectedVideos: {
-        positionTwo,
-      },
       sidebarSelectedHeroVideo,
     },
     dispatch,
@@ -78,44 +70,6 @@ const DialogWrapper = () => {
     );
   };
 
-  /**
-   * Handles the save action when a user clicks the save button.
-   */
-  const handleSave = async () => {
-    dispatch({ type: 'SET_LOADING_STATE' });
-
-    // Note: Loading state is cleared within the `saveState` service.
-    const response = await saveState(state, getPostAttribute('id'));
-    const {
-      hasError,
-      data,
-      error: {
-        message,
-      } = {},
-    } = response;
-
-    if (!hasError) {
-      // Embed id.
-      const positionTwoEmbedId = data?.embeds?.positionTwo?.id;
-
-      // Insert a new Oovvuu embed to the editor.
-      insertEmbed(positionTwoEmbedId, positionTwo, isPositionTwoEnabled);
-
-      /**
-       * saveState() returns updated state, with flag that data has been loaded
-       *   from meta. This needs to be sync'd back to state.
-       */
-      dispatch({ type: 'RESET_STATE', payload: data });
-
-      // Close the Dialog.
-      closeDialog();
-    } else {
-      dispatch({ type: 'CLEAR_LOADING_STATE' });
-
-      displayDismissableAlert({ message });
-    }
-  };
-
   return (
     <>
       <ActionButton
@@ -139,15 +93,6 @@ const DialogWrapper = () => {
           <h2 className={styles.postTitle}>
             {getPostAttribute('title')}
           </h2>
-
-          <ActionButton
-            className={styles.saveButton}
-            buttonStyle="primary"
-            onClickHandler={handleSave}
-          >
-            <SaveSVG />
-            <>{__('Save and Close', 'oovvuu')}</>
-          </ActionButton>
         </header>
         <KeywordPanel />
         <TopicsPanelWrapper />
